backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
SQLAlchemy==2.0.23
tensorflow==2.15.0
Pillow==10.1.0
reportlab==4.0.7
python-dotenv==1.0.0
Werkzeug==3.0.1
pymongo==4.5.0
